Update names and files for main dashboard view
