+++ src/client/components/Main/Main.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 import NavBar from './NavBar';
 import SideNavBar from './SideNavBar';
-import PatientProfile from './PatientProfile/PatientProfile';
 
 const Main = ({ children }) => (
   <div>
