--- src/client/components/Main/PatientProfile/PatientDetailedInfo.js
@@ -1,23 +1,0 @@
-import React, { Component } from 'react';
-
-
-const PatientDetailedInfo = () => (
-  <div className="panel panel-default patient-detailed-container">
-    <div className="patient-detailed">
-      <p>test</p>
-      <p>test</p>
-      <p>test</p>
-      <p>test</p>
-      <p>test</p>
-    </div>
-    <div className="patient-detailed">
-      <p>test</p>
-      <p>test</p>
-      <p>test</p>
-      <p>test</p>
-      <p>test</p>
-    </div>
-  </div>
-);
-
-export default PatientDetailedInfo;
