+++ src/client/components/Main/Profile/Info/DetailedInfo.js
@@ -1,0 +1,25 @@
+// TODO: check if needed
+// DELETE this component if we won't need it
+import React, { Component } from 'react';
+
+
+const DetailedInfo = () => (
+  <div className="panel panel-default person-detailed-container">
+    <div className="person-detailed">
+      <p>test</p>
+      <p>test</p>
+      <p>test</p>
+      <p>test</p>
+      <p>test</p>
+    </div>
+    <div className="person-detailed">
+      <p>test</p>
+      <p>test</p>
+      <p>test</p>
+      <p>test</p>
+      <p>test</p>
+    </div>
+  </div>
+);
+
+export default DetailedInfo;
